+++ packages/staker/src/cli.js
@@ -30,8 +30,9 @@ import {skillsGas} from './eth/saihubot-cli-skill-ethgas';
 import {skillGetBalance, skillGetEthBalance, skillGetTokensPrice} from './saihubot-cli-skill-balance';
 import {skills as skillsXDAI} from './xdai/saihubot-cli-skill-xdai';
 import {skills as skillsBSC} from './bsc/saihubot-cli-skill-bsc';
-import {skills as skillsMatic} from './matic/saihubot-cli-skill-matic';
 import {skills as skillsHeco} from './heco/saihubot-cli-skill-heco';
+import {skills as skillsMatic} from './matic/saihubot-cli-skill-matic';
+import {skills as skillOKExChain} from './okexchain/saihubot-cli-skill-okexchain';
 import {skills as skillsZkSync} from './zksync/saihubot-cli-skill-zksync';
 import {skills as skillsOptimism} from './optimism/saihubot-cli-skill-optimism';
 import {skills as skillsChain} from './saihubot-cli-skill-chainid';
@@ -83,8 +84,9 @@ const bot = new SaihuBot({
     skillFeesWtf,
     ...skillsChain,
     ...skillsBSC,
-    ...skillsMatic,
     ...skillsHeco,
+    ...skillsMatic,
+    ...skillOKExChain,
     ...skillsXDAI,
     ...skillsZkSync,
     ...skillsOptimism,